fix(D2/D6): guard transcript Kind exhaustiveness + split BuiltInChipProviders
D2: TranscriptSerializer guard replaced from hardcoded `kindInt > 2` with
`!Enum.IsDefined(typeof(TranscriptEntry.Kind), kindInt)` — self-updating,
no manual bump needed when Kind enum grows. Added default case with
LogWarning in ChatTranscript.RestoreFromReload switch so unhandled Kinds
surface immediately during manual QA. Tests 14b and 14c added:
14b (guard exhaustiveness) went RED when Kind.Task=3 was added without
fixing the guard, GREEN after the fix; went RED again on revert (both
outputs captured). 14c verifies no exception for any declared Kind.

D6: BuiltInChipProviders.cs (431 lines) split into 5 files — zero logic
changes, existing tests untouched and green. New files: AssetChipProviderBase,
HierarchyChipProvider, SceneChipProvider, ImageChipProvider (all with .meta).
Remaining 9 thin providers stay in BuiltInChipProviders.cs (151 lines).
