+++ unity-plugin/Editor/Chat/CLI/AssetChipProviderBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 034cc71a1b8749529efc9af3e43a605f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
