+++ unity-plugin/Editor/Chat/CLI/ImageChipProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 704d33ea95c04e4eaef30f2b03534a86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
