+++ unity-plugin/Editor/Chat/CLI/SceneChipProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edef2aa9daac4d0dbe9cfd8efbd7cd8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
